Split up multidimensional Stieltjes
The switch from analytical to approximation in three terms recursion formula is triggered if any dimensional component lacks analytical TTR support.
But if there are many distributions, and only a few of them needs to be approximated, the problem is still approximated in full.

This adds a fix to create TTR for each dimensions independently.
This allows the analytical and approximation values are done separately.
In addition, since the approximation is done per dimension, integration size is kept small.