Bugfix: rounding of discrete distribution (#305)
* Bugfix: rounding of discrete distribution

Rounding at 0.5 is a touchy subject as the rules differs between
implementation. The wrong assumption makes the edge cases for discrete
distributions a bit problematic as NN.5 numbers are used to construct
piecewise linear distributions. To fix this at the limits a nudge is
added to keep the rounding correct.

* Bugfix: mixing of discrete/continuous vars